rephist: add PrivCount lines to extra-info documents
Add PrivCount check and bandwidth counters to extra-info documents.
We can use these plain-text counters to test and debug PrivCount, and
check the aggregated results.

It is safe to publish these counters, because the information is already
public. (The bandwidth counter is already published elsewhere in the
extra-info document.)

Implements parts of 29004 and 29005.